Move plan fields to business table
