+++ database/factories/LeadFactory.php
@@ -25,9 +25,6 @@ class LeadFactory extends Factory
                 'last_name' => $this->faker->lastName,
                 'position' => $this->faker->word,
                 'phone_number' => $this->faker->phoneNumber,
-                'current_plan_under_cancellation' => $this->faker->boolean,
-                'current_plan_expires_at' => $this->faker->date,
-                'past_comp_claims' => $this->faker->paragraph,
             ];
         });
     }
