+++ src/Livewire/OnboardingForm/CompInsurance.php
@@ -20,7 +20,7 @@ class CompInsurance extends OnboardingFormStepComponent
 
     public function processLead(array $data)
     {
-        $this->lead->update($data);
+        $this->lead->business->update($data);
 
         return $this->lead;
     }
