+++ src/Models/LeadBusiness.php
@@ -9,6 +9,11 @@ class LeadBusiness extends BaseModel
 {
     use HasPackageFactory;
 
+    protected $casts = [
+        'current_plan_expires_at' => 'date',
+        'current_plan_under_cancellation' => 'boolean',
+    ];
+
     public function lead()
     {
         return $this->belongsTo(Lead::class);
