+++ tests/Unit/Models/LeadBusinessTest.php
@@ -4,6 +4,7 @@ namespace Roberts\Leads\Tests\Unit\Models;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Illuminate\Foundation\Testing\WithFaker;
+use Illuminate\Support\Carbon;
 use Roberts\Leads\Models\Lead;
 use Roberts\Leads\Models\LeadBusiness;
 use Roberts\Leads\Tests\TestCase;
@@ -56,6 +57,32 @@ class LeadBusinessTest extends TestCase
         $business = LeadBusiness::factory()->create(['legal_entity_type' => $type]);
 
         $this->assertEquals($type, $business->legal_entity_type);
+    }
+
+    /** @test */
+    public function it_has_the_expiration_date_for_the_current_comp_plan()
+    {
+        $business = LeadBusiness::factory()->create(['current_plan_expires_at' => $this->faker->date]);
+
+        $this->assertInstanceOf(Carbon::class, $business->current_plan_expires_at);
+    }
+
+    /** @test */
+    public function it_has_the_details_of_the_past_comp_claims()
+    {
+        $pastCompClaims = $this->faker->paragraph;
+        $business = LeadBusiness::factory()->create(['past_comp_claims' => $pastCompClaims]);
+
+        $this->assertEquals($pastCompClaims, $business->past_comp_claims);
+    }
+
+    /** @test */
+    public function it_has_an_under_cancellation_flag_for_the_current_plan()
+    {
+        $underCancellation = $this->faker->boolean;
+        $business = LeadBusiness::factory()->create(['current_plan_under_cancellation' => $underCancellation]);
+
+        $this->assertEquals($underCancellation, $business->current_plan_under_cancellation);
     }
 
     /** @test */
